+++ app/engines/rsync.sh
@@ -28,7 +28,7 @@ if [[ "${BASH_SOURCE[0]}" != "${0}" ]] ; then
     log $identifier "Rsync will be run with requested flags [$RSYNC_FLAGS]";
   fi
 
-  if env_var_is_set ARCHIVE_BACKUP; then
+  if env_var_is_set ARCHIVE_BACKUP && [ "$ARCHIVE_BACKUP" = "true" ]; then
     log $identifier "Backup will be archived";
   fi
 
@@ -43,25 +43,32 @@ current_date=$(date +%Y_%m_%d)
 target_backup_dir="$BACKUP_PATH/$current_date"
 mkdir -p "$target_backup_dir";
 
+failed=false;
+
 if /usr/bin/rsync \
     $RSYNC_FLAGS --exclude-from=/rsync-exclude \
     -e "ssh -o StrictHostKeyChecking=no -i $SSH_KEY" \
     "$SSH_USERNAME@$HOST:$REMOTE_PATH/" "$target_backup_dir/"; then
   log $identifier "Transferred successfully"
 
-  if [ "$ARCHIVE_BACKUP" = "true" ]; then
+  if env_var_is_set ARCHIVE_BACKUP && [ "$ARCHIVE_BACKUP" = "true" ]; then
+    log $identifier "Archiving backup..."
     if tar -czf "$target_backup_dir.tar.gz" -C "$target_backup_dir" .; then
       rm -rf "$target_backup_dir"
-      log $identifier "Backup archived and original backup directory removed"
+      log $identifier "Backup archived"
     else
-      log $identifier "Error in archiving process"
-      kill_container;
+      log $identifier "Error occurred while archiving backup, original backup directory will persist"
+      failed=true;
     fi
   fi
 
 else
   log $identifier "Some errors occurred while transferring from $SSH_USERNAME@$HOST:$REMOTE_PATH to $target_backup_dir
                    key:[$SSH_KEY] exclude:[$(cat /rsync-exclude)]"
+  failed=true;
+fi
+
+if [ $failed = true ]; then
   log $identifier "Killing container due to errors";
   kill_container;
 fi
